Добавлена функция GenGameName

+++ DZ_08.11.25/Game.cs
@@ -10,7 +10,7 @@ namespace DZ_08._11._25
     {
         public string Name;
         public int Downlouds;
-        public int Rates;
+        public double Rates;
 
     }
 }
